refactor: eliminar interfaces UseCase, settings endpoint y fixes PDF/solicitudes
Refactorizacion del modulo auth:
- Elimina las 5 interfaces UseCase (LoginUseCase, CreateUserUseCase, etc.)
- Controllers inyectan UserService directamente — mas simple y legible
- UserService expone metodos concretos: login(), create(), changePassword(), etc.

Nuevo endpoint GET /api/v1/settings:
- Expone datos publicos de la empresa (companyName, authorizers)
- Lista de autorizadores configurable via COMPANY_AUTHORIZERS env var
- Disponible sin autenticacion para el frontend

Fix PDF (500 Internal Server Error):
- Elimina java.math.BigDecimal.ZERO de expresion Thymeleaf (no es accesible desde Spring EL)

Fix V5 migracion:
- Cambia http_status de SMALLINT a INTEGER para coincidir con la entidad JPA

CompanyProperties:
- Agrega campo authorizers (List<String>) para los selects del frontend

+++ src/main/resources/db/migration/V5__idempotency_and_documents.sql
@@ -9,7 +9,7 @@ CREATE TABLE idempotency_keys (
     id           UUID         PRIMARY KEY,
     key_hash     VARCHAR(64)  NOT NULL UNIQUE,   -- SHA-256 de "userId:key"
     endpoint     VARCHAR(200) NOT NULL,
-    http_status  SMALLINT     NOT NULL,
+    http_status  INTEGER      NOT NULL,
     response     TEXT         NOT NULL,          -- JSON de la respuesta
     created_at   TIMESTAMPTZ  NOT NULL DEFAULT NOW(),
     expires_at   TIMESTAMPTZ  NOT NULL           -- TTL: 24h por defecto
